Make Radio unbundled - part 2/2
This CL makes the RadioAppService run in the default process so that
we can use a simple bound service instead of the previous aidl connection
as the aidl framework files used by radio are not available in an unbundled
build.

Also replaces all the uses of private resources.
And changes the make file to compile as an unbundled app \o/

Bug: 319935691
Test: manual on seahawk both with the CarRadioApp doing the UI and then
      with CarMediaApp doing the UI (disabling the exception to open the
      radio activity).

Change-Id: I1a824e8460015b2ea88a9ca64a1fcd4b1ce1a64b