Set Java and Kotlin target to JDK 8
This commit sets the Java and Kotlin target to JDK 8 for all projects.
It introduces a new Gra
dle plugin `buildSrc.jdk.jdk8` to manage this configuration. Versions of the plugin for JDK 11 and 17 are also included for future apps.
This change ensures compatibility and consistency across the codebase and is required to support Gradle 8+

Bug: b/356141874
Test: ./gradlew assemble assembleAndroidTest test
Change-Id: Idf3c24c7e0da8e4efe1c95c1db6da8ad5c3df641